# ECommerceApplication/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Ecommerce API
  description: API documentation for Ecommerce system (admin and user roles)
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api
    description: Local server
...
paths:
...
  /users:
    get:
      tags: [User]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      # Role: ADMIN
  /users/admin/{id}:
    get:
      tags: [User]
      summary: Get user by ID (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      # Role: ADMIN

    delete:
      tags: [User]
      summary: Delete user by ID (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted
      # Role: ADMIN
  /users/me:
    get:
      tags: [User]
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      # Role: USER

    put:
      tags: [User]
      summary: Update current user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Updated user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      # Role: USER
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        roles:
          type: array
          items:
            type: string
        # Add other user fields as per your entity
      required:
        - email
        - password
...